# app/routers/factory.py
"""
Programmatic Content Factory Router
-------------------------------------
Turn one source asset into dozens of format-adapted content pieces.
Then let the algorithmic scheduler pick optimal posting windows.

Week 5: Distribution at superhuman speed.
Week 6+: Live Campaign Mode — hook-based 1-to-20 multiplication
with 9:16 vertical rendering and animated captions.
"""

from fastapi import APIRouter, Depends, HTTPException, Query, status
from datetime import datetime, timezone

from ..core.auth import verify_api_key
from ..core.dependencies import get_content_factory
from ..services.content_factory import ContentFactory
from ..schemas.content_factory import (
    ContentPipelineRequest,
    ContentPipelineResponse,
    GeneratedContent,
    ContentListResponse,
    AnalyticsIngestRequest,
    AnalyticsIngestResponse,
    ScheduleRequest,
    ScheduleResponse,
    LiveCampaignRequest,
    LiveCampaignResponse,
)

router = APIRouter(
    prefix="/v1/factory",
    tags=["Content Factory & Scheduling"],
    responses={
        401: {"description": "Missing API key"},
        403: {"description": "Invalid API key or access denied"},
    },
)


# --- Content Pipeline ---

@router.post(
    "/pipelines",
    response_model=ContentPipelineResponse,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Create a content generation pipeline",
    description=(
        "Submit a source asset and target formats. The factory will "
        "generate multiple content pieces adapted to each format: "
        "short videos, static images, text posts, audiograms, carousels, "
        "blog excerpts, and email snippets. Processing is async."
    ),
)
async def create_pipeline(
    request: ContentPipelineRequest,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    pipeline = await factory.create_pipeline(
        title=request.title,
        target_formats=request.target_formats,
        source_clip_id=request.source_clip_id,
        source_url=request.source_url,
        brand_config=request.brand_config,
        language=request.language,
        auto_schedule=request.auto_schedule,
        owner_key=api_key,
    )

    return ContentPipelineResponse(
        pipeline_id=pipeline.pipeline_id,
        title=pipeline.title,
        source_type="clip" if pipeline.source_clip_id else "url" if pipeline.source_url else "none",
        target_formats=pipeline.target_formats,
        status=pipeline.status,
        estimated_pieces=factory.estimate_pieces(pipeline.target_formats),
    )


@router.get(
    "/pipelines/{pipeline_id}",
    summary="Get pipeline status",
    description="Check the status of a content generation pipeline.",
)
async def get_pipeline(
    pipeline_id: str,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    pipeline = await factory.store.get_pipeline(pipeline_id)
    if not pipeline:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail={"error": "pipeline_not_found"},
        )
    return {
        "pipeline_id": pipeline.pipeline_id,
        "title": pipeline.title,
        "status": pipeline.status,
        "target_formats": [f.value for f in pipeline.target_formats],
        "pieces_generated": len(pipeline.content_pieces),
        "created_at": pipeline.created_at.isoformat(),
    }


@router.get(
    "/pipelines/{pipeline_id}/content",
    response_model=ContentListResponse,
    summary="List generated content pieces",
    description="Retrieve all content pieces generated by a pipeline.",
)
async def list_pipeline_content(
    pipeline_id: str,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    pipeline = await factory.store.get_pipeline(pipeline_id)
    if not pipeline:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail={"error": "pipeline_not_found"},
        )

    content = await factory.list_pipeline_content(pipeline_id)
    return ContentListResponse(
        content=content,
        total=len(content),
        pipeline_id=pipeline_id,
    )


@router.get(
    "/content/{content_id}",
    response_model=GeneratedContent,
    summary="Get content piece details",
    description="Retrieve metadata and download URL for a specific content piece.",
)
async def get_content(
    content_id: str,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    content = await factory.get_content(content_id)
    if not content:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail={"error": "content_not_found"},
        )
    return content


# --- Live Campaign Mode ---

@router.post(
    "/campaigns",
    response_model=LiveCampaignResponse,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Launch a live content campaign",
    description=(
        "The 'Big Red Button' — submit a source video URL and targeted hooks. "
        "Each hook is a specific segment (e.g., 30-45s reaction clip, 60s explainer). "
        "The factory applies the 1-to-N multiplication rule: each hook produces "
        "multiple format-adapted pieces (short videos, quote cards, carousels, text posts) "
        "all rendered in 9:16 vertical with animated captions. "
        "Then the AlgorithmicScheduler staggers posts across platforms "
        "(TikTok, YouTube Shorts, Instagram Reels) to maximize view velocity."
    ),
)
async def launch_campaign(
    request: LiveCampaignRequest,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    result = await factory.launch_campaign(
        campaign_title=request.campaign_title,
        source_url=request.source_url,
        hooks=request.hooks,
        brand_config=request.brand_config,
        caption_style=request.caption_style,
        aspect_ratio=request.aspect_ratio,
        platforms=request.platforms,
        max_posts_per_day=request.max_posts_per_day,
        language=request.language,
        auto_schedule=request.auto_schedule,
        owner_key=api_key,
    )
    return result


@router.get(
    "/campaigns/{campaign_id}",
    summary="Get campaign status",
    description="Check the status and results of a live content campaign.",
)
async def get_campaign(
    campaign_id: str,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    campaign = await factory.store.get_campaign(campaign_id)
    if not campaign:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail={"error": "campaign_not_found"},
        )
    return {
        "campaign_id": campaign.campaign_id,
        "campaign_title": campaign.campaign_title,
        "source_url": campaign.source_url,
        "status": campaign.status,
        "hooks_count": len(campaign.hooks),
        "pipeline_ids": campaign.pipeline_ids,
        "created_at": campaign.created_at.isoformat(),
    }


@router.get(
    "/campaigns",
    summary="List all campaigns",
    description="Retrieve all live content campaigns.",
)
async def list_campaigns(
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    campaigns = await factory.store.list_campaigns()
    return {
        "campaigns": [
            {
                "campaign_id": c.campaign_id,
                "campaign_title": c.campaign_title,
                "status": c.status,
                "hooks_count": len(c.hooks),
                "created_at": c.created_at.isoformat(),
            }
            for c in campaigns
        ],
        "total": len(campaigns),
    }


# --- Algorithmic Scheduling ---

@router.post(
    "/analytics",
    response_model=AnalyticsIngestResponse,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Ingest engagement analytics",
    description=(
        "Feed engagement data from platform APIs into the scheduling engine. "
        "The scheduler uses this data to learn optimal posting windows. "
        "More data = smarter scheduling = higher view velocity."
    ),
)
async def ingest_analytics(
    request: AnalyticsIngestRequest,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    summary = await factory.scheduler.ingest_analytics(request.metrics)
    return AnalyticsIngestResponse(
        ingested=len(request.metrics),
        platform_summary=summary,
    )


@router.get(
    "/analytics/summary",
    summary="Get analytics summary",
    description="View summary of ingested engagement data.",
)
async def get_analytics_summary(
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    return await factory.scheduler.get_analytics_summary()


@router.post(
    "/schedule",
    response_model=ScheduleResponse,
    summary="Get optimal posting schedule",
    description=(
        "Generate an algorithmically-optimized posting schedule for content pieces. "
        "The scheduler analyzes historical engagement data per platform to pick "
        "the highest-engagement time slots. Posts are spread across days to avoid "
        "audience fatigue (configurable via max_posts_per_day)."
    ),
)
async def get_schedule(
    request: ScheduleRequest,
    api_key: str = Depends(verify_api_key),
    factory: ContentFactory = Depends(get_content_factory),
):
    recommendations = await factory.scheduler.recommend(
        content_ids=request.content_ids,
        platforms=request.platforms,
        earliest=request.earliest,
        latest=request.latest,
        max_per_day=request.max_posts_per_day,
    )

    # Build date range string
    if recommendations:
        start = recommendations[0].recommended_time.strftime("%Y-%m-%d")
        end = recommendations[-1].recommended_time.strftime("%Y-%m-%d")
        date_range = f"{start} to {end}"
    else:
        date_range = "no recommendations"

    return ScheduleResponse(
        recommendations=recommendations,
        total_scheduled=len(recommendations),
        date_range=date_range,
    )