# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent-Native Middleware API",
    "description": "## Agent-Native Middleware API\n\nHeadless infrastructure for the B2A (Business-to-Agent) economy. This API provides four core services that give AI agents the 'hands' and 'eyes' to manipulate the digital and physical world.\n\n### Service Pillars\n\n- **IoT Protocol Bridge** (`/v1/iot`) \u2014 Secure, ACL-enforced protocol translation for IoT devices (MQTT, CoAP, Zigbee, etc.)\n- **Autonomous Product Manager** (`/v1/telemetry`) \u2014 Ingest telemetry, detect anomalies, auto-generate pull requests to fix bugs\n- **Programmatic Media Engine** (`/v1/media`) \u2014 Video ingestion, viral hook detection, reframing, captioning, and cross-platform distribution\n- **Agent Communications** (`/v1/comms`) \u2014 Agent registration, structured messaging, capability discovery, and swarm task handoffs\n\n### Authentication\n\nAll endpoints require an API key passed via the `X-API-Key` header.\n\n### For Agents\n\nThis API is designed for programmatic consumption. See `/llm.txt` for LLM-optimized documentation and `/openapi.json` for the full spec.",
    "contact": {
      "name": "Agent-Native Middleware",
      "email": "api@yourdomain.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.yourdomain.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
...
    }
  ],
  "paths": {
...
    "/v1/factory/pipelines": {
      "post": {
        "tags": [
          "Content Factory & Scheduling"
        ],
        "summary": "Create a content generation pipeline",
        "description": "Submit a source asset and target formats. The factory will generate multiple content pieces adapted to each format: short videos, static images, text posts, audiograms, carousels, blog excerpts, and email snippets. Processing is async.",
        "operationId": "create_pipeline_v1_factory_pipelines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentPipelineResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key or access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/factory/pipelines/{pipeline_id}": {
      "get": {
        "tags": [
          "Content Factory & Scheduling"
        ],
        "summary": "Get pipeline status",
        "description": "Check the status of a content generation pipeline.",
        "operationId": "get_pipeline_v1_factory_pipelines__pipeline_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key or access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/factory/pipelines/{pipeline_id}/content": {
      "get": {
        "tags": [
          "Content Factory & Scheduling"
        ],
        "summary": "List generated content pieces",
        "description": "Retrieve all content pieces generated by a pipeline.",
        "operationId": "list_pipeline_content_v1_factory_pipelines__pipeline_id__content_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key or access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/factory/content/{content_id}": {
      "get": {
        "tags": [
          "Content Factory & Scheduling"
        ],
        "summary": "Get content piece details",
        "description": "Retrieve metadata and download URL for a specific content piece.",
        "operationId": "get_content_v1_factory_content__content_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedContent"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key or access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/factory/analytics": {
      "post": {
        "tags": [
          "Content Factory & Scheduling"
        ],
        "summary": "Ingest engagement analytics",
        "description": "Feed engagement data from platform APIs into the scheduling engine. The scheduler uses this data to learn optimal posting windows. More data = smarter scheduling = higher view velocity.",
        "operationId": "ingest_analytics_v1_factory_analytics_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsIngestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key or access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/factory/analytics/summary": {
      "get": {
        "tags": [
          "Content Factory & Scheduling"
        ],
        "summary": "Get analytics summary",
        "description": "View summary of ingested engagement data.",
        "operationId": "get_analytics_summary_v1_factory_analytics_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key or access denied"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/factory/schedule": {
      "post": {
        "tags": [
          "Content Factory & Scheduling"
        ],
        "summary": "Get optimal posting schedule",
        "description": "Generate an algorithmically-optimized posting schedule for content pieces. The scheduler analyzes historical engagement data per platform to pick the highest-engagement time slots. Posts are spread across days to avoid audience fatigue (configurable via max_posts_per_day).",
        "operationId": "get_schedule_v1_factory_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key or access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyticsIngestRequest": {
        "properties": {
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/PlatformAnalytics"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Metrics",
            "description": "Engagement metrics from platform APIs. The scheduler learns from this data."
          }
        },
        "type": "object",
        "required": [
          "metrics"
        ],
        "title": "AnalyticsIngestRequest",
        "description": "Submit engagement data to improve scheduling."
      },
      "AnalyticsIngestResponse": {
        "properties": {
          "ingested": {
            "type": "integer",
            "title": "Ingested"
          },
          "platform_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Platform Summary",
            "description": "Per-platform ingestion counts."
          }
        },
        "type": "object",
        "required": [
          "ingested",
          "platform_summary"
        ],
        "title": "AnalyticsIngestResponse"
      },
...
      "ContentFormat": {
        "type": "string",
        "enum": [
          "short_video",
          "long_video",
          "audiogram",
          "carousel",
          "static_image",
          "text_post",
          "blog_excerpt",
          "email_snippet"
        ],
        "title": "ContentFormat",
        "description": "Output content formats."
      },
      "ContentListResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/GeneratedContent"
            },
            "type": "array",
            "title": "Content"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          }
        },
        "type": "object",
        "required": [
          "content",
          "total",
          "pipeline_id"
        ],
        "title": "ContentListResponse",
        "description": "Paginated content listing."
      },
      "ContentPipelineRequest": {
        "properties": {
          "source_clip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Clip Id",
            "description": "ID of an existing clip from the Media Engine. Mutually exclusive with source_url."
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url",
            "description": "URL of raw source content (video, audio, image)."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Working title for the content batch.",
            "examples": [
              "Q1 Product Launch Highlights"
            ]
          },
          "target_formats": {
            "items": {
              "$ref": "#/components/schemas/ContentFormat"
            },
            "type": "array",
            "title": "Target Formats",
            "description": "Output formats to generate. Each source becomes multiple format-adapted pieces.",
            "default": [
              "short_video",
              "static_image",
              "text_post"
            ]
          },
          "brand_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Brand Config",
            "description": "Brand customization: colors, fonts, logo_url, watermark settings.",
            "examples": [
              {
                "font_family": "Inter",
                "logo_url": "https://cdn.example.com/logo.png",
                "primary_color": "#FF6B00",
                "watermark_position": "bottom_right"
              }
            ]
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "auto_schedule": {
            "type": "boolean",
            "title": "Auto Schedule",
            "description": "Automatically schedule distribution using the algorithmic scheduler.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ContentPipelineRequest",
        "description": "Submit a source asset for multi-format content generation."
      },
      "ContentPipelineResponse": {
        "properties": {
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "target_formats": {
            "items": {
              "$ref": "#/components/schemas/ContentFormat"
            },
            "type": "array",
            "title": "Target Formats"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "estimated_pieces": {
            "type": "integer",
            "title": "Estimated Pieces",
            "description": "Estimated number of content pieces that will be generated."
          }
        },
        "type": "object",
        "required": [
          "pipeline_id",
          "title",
          "source_type",
          "target_formats",
          "status",
          "estimated_pieces"
        ],
        "title": "ContentPipelineResponse",
        "description": "Response after initiating a content pipeline."
      },
      "ContentStatus": {
        "type": "string",
        "enum": [
          "queued",
          "rendering",
          "ready",
          "distributed",
          "failed"
        ],
        "title": "ContentStatus"
      },
...
      "GeneratedContent": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          },
          "format": {
            "$ref": "#/components/schemas/ContentFormat"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "dimensions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions",
            "description": "WxH for image/video formats.",
            "examples": [
              "1080x1920"
            ]
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "pipeline_id",
          "format",
          "title",
          "download_url",
          "status",
          "generated_at"
        ],
        "title": "GeneratedContent",
        "description": "A single generated content piece."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlatformAnalytics": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "metric_type": {
            "$ref": "#/components/schemas/EngagementMetricType"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "metric_type",
          "value",
          "recorded_at"
        ],
        "title": "PlatformAnalytics",
        "description": "Engagement analytics for a specific platform."
      },
...
      "ScheduleRequest": {
        "properties": {
          "content_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Ids",
            "description": "Content piece IDs to schedule."
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Platforms",
            "description": "Target platforms."
          },
          "earliest": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest",
            "description": "Earliest acceptable posting time. Defaults to now."
          },
          "latest": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest",
            "description": "Latest acceptable posting time. Defaults to 7 days from now."
          },
          "max_posts_per_day": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Posts Per Day",
            "description": "Maximum posts per platform per day to avoid audience fatigue.",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "content_ids",
          "platforms"
        ],
        "title": "ScheduleRequest",
        "description": "Request optimal scheduling for content pieces."
      },
      "ScheduleResponse": {
        "properties": {
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/ScheduleRecommendation"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "total_scheduled": {
            "type": "integer",
            "title": "Total Scheduled"
          },
          "date_range": {
            "type": "string",
            "title": "Date Range"
          }
        },
        "type": "object",
        "required": [
          "recommendations",
          "total_scheduled",
          "date_range"
        ],
        "title": "ScheduleResponse",
        "description": "Full schedule with per-content, per-platform recommendations."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API key for agent authentication. Pass in the X-API-Key header.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}